+++ Src/Roflcopter.Plugin.Tests/test/data/UnitTesting/ConvertToParameterizedTestContextActionTest/Availability.cs
@@ -1,0 +1,26 @@
+﻿using NUnit.Framework;
+
+namespace Roflcopter.Sample.UnitTesting.ConvertToParameterizedTestContextActionTest
+{
+    public class Availability
+    {
+        {off}
+
+        public void NoTestMethod()
+        {
+            {off}
+        }
+
+        {on}[Test]
+        public void TestMethod()
+        {
+            {on}
+        }{on}
+    
+        [Test]
+        public void TestMethodWithParameter(string param)
+        {
+            {off}
+        }
+    }
+}