+++ Src/Roflcopter.Plugin.Tests/test/data/UnitTesting/ConvertToParameterizedTestContextActionTest/TestMethodWithoutParameters.cs
@@ -1,0 +1,13 @@
+﻿using NUnit.Framework;
+
+namespace Roflcopter.Sample.UnitTesting.ConvertToParameterizedTestContextActionTest
+{
+    public class TestMethodWithoutParameters
+    {
+        [Test]
+        public void TestMethod()
+        {
+            {caret}
+        }
+    }
+}